semana 5/requirements.txt
# Dependencias existentes
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23

# Nuevas dependencias para autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6